Tweak zero results page a big
Further refinement of #2067, some notes there.

Really wanted to get rid of the extra space at the top before the text -- but blacklight made it hard. That extra space is margin and padding in the "sort and per page" element, which is empty when we have zero results -- after we removed the extratenous "no results found" message. But there was no good way to get rid of that emtpy struture sitting there, with it's margin and padding, without LOTS of deep overriding of blacklight, which did not seem worth it. So I guess we'll live with the extra whitespace on top, although it's pretty ugly.