hackily explicitly start Rails.ujs only if it hasn't already been started
Using possibly internal hook to discover if it's been started. Not sure why it's without this being started in production but not in dev; instead of understanding the root cause, we hackily resort to starting it only conditionally, to avoid *double* starting it in production resulting in an error; or not starting it at all in dev.