Change public label of link category to 'Scientific Biography'
From previously 'Historical biography'.  (case was inconsistent with other labels).

Per #1867

Note that internal data was and remains `institute_biography` (and staff metadata form was and remains 'Institute Biography'. We just had an alternate label for public view, already. So we can change this easily without a data migration, hooray.