Fix edit collection test for navigation
This test was testing the wrong thing, and passing accidentally -- we used to return to admin collection list page after submitting collection edit form, but a while ago changed to go to public show page. The test was accidentally passing on matching a different h1 that said 'Digital Collections'... which broke in my header refator, so I noticed the error. Let's fix on master.