Make sort widget line up flush with right margin
By overriding some styles that came from Blacklight's use of bootstrap, in a case where there were more things on screen I guess, or just a different layout. We don't want it.

This makes the right edge of the sort dropdown be nice and flush with other content right edges.