Replace btn-outline-primary with btn-outline-secondary in admin screens
In new redesign colors, after resetting bootstrap colors to what made sense for our uses: btn-outline-primary is now a teal outline;  btn-outline-secondary is a black outline. The black outline is actually in new site styles, and looks good. The teal outline is not, and does not look as good.  More consistency with new site styles, and looks better.

These are all in admin screens.