revert muted/greyed out text to pre-redesign color
It goes fine with palette, and I like it better than the new one I was trying (which had a blue tint to go with new brand greys). This one is better.

Note the new logo itself uses a grey for sub-head "Museum & Library" that isn't actually in Brand Guide pallette either, so there's precedent. (Yeah, I tried identifying and using that one too, too light for us)