update video.js from 7.x to 8.x
skip ahead/forward by X seconds is functionality now built into 8.x -- we were formerly using plugin `videojs-seek-buttons`, but it's README now recommends:

> However Video.js 8.2.0 + has a built-in seek buttons functionality. Consider using that instead of this plugin.

So we removed that plugin, and switched config over to use built-in functionality. We were previusly skipping forward 30 and back 15. The built-in func only supports skipping by 5,10, or 30 seconds -- so the skip back is now 10.

The built-in functionality shows the seconds you will be skipping on screen though, when before it did not. So that's an improvement, that had been requested by some of our local stakeholders too I recall.