Tweak config of mem_cache_store used for rack-attack
With context from the memcachier heroku docs. We are attempting to reduce/eliminate all the errors we get for memcached being unavailable.

We add a 'pool', recommended there for use with multi-thread puma, which we are using.

We also, while looking at the config, go back on our decision to eliminate retry after error, instead enabling a pretty small 80ms retry delay.

Ref #2143, trying harder with memcache before giving up