Rake task to set flag and add OCR to existing works matching spec
Spec by Annabel in #2080

Fetches items matching spec, sets ocr_requested flag, queues up background job (on special_jobs queue) to do actual OCR.

I have run this in development to make sure it seems to work. We don't really want to run it in staging, because we don't want to spend that money/CPU time on OCR'ing everything in staging, we want to do it in production. So I think we'll just have to try.