Allow sync_prod_to_staging to optionally not put staging app into maintenance mode
In our actual work, it turns out, we'd often rather have a few hours of staging being inconsistent with some files missing, then a few hours of staging outage.

This is still by default maintenance mode, you have to choose to do it the way that will be less stable.