Remove and stop using download_small derivatives
In public UI, call the download_medium `small` -- if there are two options to users, it should be `small` and `large`, not `medium` and `larger`.

But it's harder to rename derivative internally, the name `download_medium` is actually part of its' path on S3. So we'll leave it "medium" internally everywhere, just change it to be labelled `small` in the single part of codebase responsible for presenting to user.