Extract logic on whether to include a PDF to OUTSIDE ImageDownloadOptions
ImageDownloadOptions just gets a boolean whether to include pdf, the caller has to decide if it wants it or not.

Is this better? Hard to say. All this dropdown logic has become very convoluted. I have not been able to figure out how to make it incontrovertibly better, and am just resigning to this being good enough. What this does do is:

* Resolve the n-query problem in #2378. I still don't totally understand what was even going on there, but this resolve it.
* Set things up for #2372 to be easier, where in some cases the dropdown menu shouldn't include the whole-work-download options at all.

Separating this out from #2372 for easier review.

I am definitely not thrilled with how weird this has all become, but best I can do for now. If you have any other ideas though, let's talk.