Dashboard observability/alert of promotion/ingest failed
Some easy simple hacky dashboard observabiilty of asets that have had promotion failed. (This is an unfortuante case of using different internal shrine language "promotion" than staff-user-facing language ("ingest"), so be it.

Yes, we potentially do a query on every single admin sidebar display to see if there are any ingest-failed. But it should be a fairly quick query -- we did ensure it's backed by an index. And actually we'll add some Rails cache -- but we haven't actually set up Rails cache, so it won't actually be used yet, but will if we ever set up a cache in production!

There is a filter for promotoin-failed assets on the admin asset list, but no UI to trigger it except clicking on the alert in sidebar.

All pretty hacky but simple enough that it doesn't seem like a big deal, and just a last resort to make sure admins notice if an asset has failed ingest, if they miss other ways to notice.