A PDF asset should not display download option for whole-work-pdf derivative
This is a band-aid on a mess -- actually worse than a band-aid, it's making the mess worse, with a quick fix. But we fix #2401 with this at least.

Long-term we need to refactor how this works. Having multiple different points that are responsible for deciding what download links to include, based on characteristics of the asset and it's parent itself -- is not working, not viable.  Left a ticket for a refactor here: #2416

This is a quick fix to our current bug, even though it makes the underlying issue of convoluted logic built of exceptions-to-exceptions in non-local places temporarily worse.