Remember Oral History request form entries to re-use
In a cookie, when same browser is used.  cookie lasts a day, but TTL is refreshed on each succesful form submission, so a day past last time you used.

With an actual system test demonstrating the user really does get a pre-filled form, since this is important behavior, that we ourselves might not use so might not notice if working or not.