move restricted derivatives to factory and stop using #dup on ActiveRecord
In Rails 7.1 it was creating problems with dup'ing the friendlier_id, I think because Rails 7.1 is better at actually retrieving the database-set friendlier_id

This was a messy way to do it anyway, dup'ing ActiveRecord objects is kind of asking for trouble. Factories are the right way to DRY a model in a particular state for test, that's what they're for.