Memoize fetch of WorkFileListShowComponent#all_members
This really should have been memoized, as @bensheldon noticed in investigating #2449 -- or actually, should it be passed in from a caller that may already have this list, isntead of being fetched again in the component anyway?  (The biggest challenge we have with components!).  Not sure how to do that, but we can memoize it, and do so here.