store selected exiftool characterization in normalized metadata location
We store complete exiftool characterization result in native json format, so we can always refer to it later, if we want more from it or there were bugs in using it. This is good archival practice.

But future exiftool versions may change their results. For anything we want to use computationally, we also want to extract and store in a standardized normalized metadata location. For now, we defintely need dpi for giving to InternetArchive BookReader ref #2463. For now only that!

with rake task to extract selected normalized metadata from exiftool characterization for existing data