viewer, escape key should close modal even when input is focused
Without this, there was a weird bug where the escape key for some reason caused the whole modal to increase in size briefly then return to size. I have NO idea what's going on there... but we'll just return to desired behavior where escape key is caught and triggers close of viewer modal.

This is all getting a bit rickety.