Pre-fill search-inside form on work with previous query you came from
Implemented clever-hackily by:

* links from search results have a hash/fragment added to them, #q={query}
* work page has javascript that notices this, uses it to pre-fill query, then REMOVES it to keep url tidy

Confirmed that google indexing seems to ignore #fragmentIdentifiers in URLs, it won't think it's a different page. https://searchengineland.com/the-right-way-to-apply-javascript-to-your-links-for-seo-333888

Code for adding/removing things from fragmentIdentifier avoids stepping on other key=value pairs that may be in there should we want to use this more in future.