Remove #q={query} from work page URL even if the work does not support search-inside
That anchor is used to pre-fill the search-inside box on the work page. When we have a search-inside box, we pre-fill from there, then remove it from the page URL to have a clean URL in address bar.

But even if we don't have a search-inside box, but the anchor is still there, we still want to clean it up and remove it.