avoid double adding OCR highlight boxes
This was triggered in certain cases where highlights were added by BOTH an openseadragon 'open' event handler, AND in getSearchResults(). Specifically, when opening a bookmarked URL to specific search results and a page with hits.

The effect was two highlight divs on one on top of the other, both semi-transparent, but effectively less transparent than intended since there were two of them.

This is all getting a bit tricky