Correct calculation of when need to scroll thumbnail list in viewer small screen vertical layout
Not sure why we had an off by one error before, and have to add one pixel adjustment. Makes no sense to me, but doens't really hurt, and keeps us from scrolling unnecessarily.

(If in some browsers it avoids scrolling when 1px of the thumbnail is off the screen, that's not really a problem, so.)