update package.json to allow node 22, which is now on my Mac
the Node version on my mac somehow became 22.1.0. Which was not allowed by our package.json. It's fairly new we have our package.json specifying bounds of allowable node versions. I wonder if we want to just let it be unbounded on the upper end after all? We'll see.

But for now, node 22 seems to work fine, update package.json to allow it.