scihist:solr_cloud:create_collection rake task can handle an already existing configset, and use that
Normally doesn't happen, but ran into that state when trying delete and re-create a collection in #2683

So this makes this task more resilient to that condition, so it does what it needs even when we're in that state trying to deal with a problem!

It's okay to assume an existing configset with same name is the same configset, because we put a digest code into config set name. If you need more precise things, like manualy deleting the config set instead of re-using it, you just gotta open a console and use SolrConfigsetUpdater helper directly.