Get sitemap URL to go through cloudfront
Sitemap is stored on derivatives, which is no longer directly accessible, we need to access it thorugh cloudfront, and tell crawlers to do such.

This is a quick and easy way to do so, just add another config variable. "Cleaner" would be using a shrine storage or something, but that would require more refacotring and you know what this is good enough.