Disable YJIT on heroku non-web (worker, console) to save RAM
The performance enhancements of YJIT are less useful and not needed on eg bg worker dynos. Where we also are running dynos that are very tight on RAM, and don't want to spend it on YJIT RAM overhead.

Heroku sets an ENV variable DYNO to eg "worker.2" or "web.1", so we can look at it, and if it set to something other than "web.X", say, nah, disable the otherwise default on YJIT.