DownloadOption can have a separate 'download_url' to force downloads
Which is used by our born-digital PDF links so that we can have a link for 'inline' viewing of the PDFs on the link itself (opening in new tab); and separate url for forced-dopwnload (with content-disposition headers) on the "download" button.

This wasn't relevant before, because the AJAXy stuff for our on-demand derivatives -- its' not feasible to have inline view links at all (or at least we haven't tackled the challenge), so it was all forced download.

But with THESE NEW links, we can provide an inline view option. (Is it bad it's not consistent? I think it's okay)