Fix collection show-with-search customization for Rails 8
The custom over-ride of the "#facet" method has become out of date in Rails 8, was calling the wrong methods, raised.

However on a close look -- I couldn't figure out what this override was for anyway, it looked exactly like current base implementation. Tried to see if things would work without the override at all. They appeared to?

I think it has been unnecessary ever since commit 6d96dd27451dc86e03 in 2019 -- prior to that we were using param keys differnetly than Blacklight so needed to override method to use our different keys, but as of that commit that was no longer true and the override was prob unnecessary since then.

With browser test to make sure facet "more" link was working -- that was what was broken.