Remove bootstrap gem from gemfile
We now use bootstrap 5 via npm with vite, not from the rubygem. The rubygem is unneeded.

Remove duplicate sprockets/sprockets-rails in gemfile, only sprockets-rails, higher in gemfile. Having them both with sprockets first was effecting load order in a weird way that prevented startup, in the absence of sassc-rails which had been a bootstrap gem indirect dependency.