Remove bootstrap gem, as we're no longer using bootstram from rubygem
Shuffle around dependencies required when bootstrap's indirect dependencies disappear. We require sprockets-rails directly, and for some reason it requires sassc-rails even though we aren't actually using it.

We are at this point not compiling/bundling anything with sprockets, it's all through vite, so removing all non-default sprockets config that doesn't apply anymore too.

we should be set up for an easy switch to rails-preferred propshaft in the future, as we are not using sprockets for bundling.