try to make copy to clipboard test more reliable
In some cases the 'copy to clipboard' button is ending up off-screen when capybara tries to click it. I think this is because bootstrap accordion is doing some crazy animation when we automatically open/close the different elements. We can try to work around it by actually scrolling the browser to right position?

a) This may be an actual problem for users of course that we should try to solve, I have kind of noticed it

b) I'm not sure if our previous attempts to improve reliability with the rescue right here are necessary working or not? But leaving them in too!