Remove require that is no longer required in Rails 8 to patch IRB how we like
It resulted in deprecation warning:

```
DEPRECATION WARNING: `rails/console/methods` has been deprecated and will be removed in Rails 8.1.
Please directly use IRB's extension API to add new commands or helpers to the console.
For more details, please visit: https://github.com/ruby/irb/blob/master/EXTEND_IRB.md
```

But we didn't need the require at all in Rails 8 to patch what we wanted to patch -- not sure if extension API would let us or be appropriate to patch what we want, but this seems to be working.