Stop logging "informational" rate limits
We are throttling at 180 reqs per minute, and log when an IP is throttled.

We also were monitoring a lower rate limit, to just log informationally. Especially as we were trying to get a sense of what was going on, that was maybe interesting.

But now that we've done it for a while, those voluminous logs aren't giving us much info, we aren't really using them. By no longer tracking them, we can increase our performance a bit (one less thing that's tracked on every request; less RAM in memcached to track it), and decrease our logging-and-alerting-stuff-we-don't-look-at.