Capybara specs should wait for page to be returned before checking for database change
Otherwise, if the app is slower than capybara, when we're checking for a db change, maybe it hasn't been done yet. We actually have to wait for the app to return the next page to ensure it finished the action before checking for DB change.

Not really sure why this has started becoming a problem recently, hopefully we haven't slowed down our app. But these are corrections to do things actually reliably correct regardless.