Drop solr timeout to 2s, with one retry after 300 ms wait
So maximum end-to-end time should be around 2000ms + 300ms + 2000ms == 4.3s

Previously maximum wait time was:

4000ms + 300ms + 4000ms + 600ms + 4000ms == 12.9 seconds.

That's far too long for a single puma thread to be held up waiting on solr, and definitely caused problems when solr was having a brief outage with all our puma threads in use.

This may still be enough time to caues problems, depending on our traffic volume. We will see.

We are having ~4 times a day where Solr is unresponsive for 8-10s, very frustrating, have to figure that one out.

Some references:

* https://github.com/sciencehistory/scihist_digicoll/issues/2684
* https://github.com/sciencehistory/scihist_digicoll/issues/2684#issuecomment-2486426237