DZI manifest file location is now tracked by a Shrine uploaded file
Stored in an attr_json, an implementation we previously were using and verified.

Before this change, the DZI file location was assumed based on current asset fingerprint checksum. That meant that it automatically changed when the asset changed -- good. But we had no way to move it to a new location _without_ changing the asset fingerprint. And since we are delivering them with cache-forever headers, when we have an improvement to the generation algorithm and want to regenerate them -- we need to move them to a new location.

Persisting current location in the db lets us change the location on demand, although of course we have to generate them at a different location, but we can change the logic for picking a location -- and the old location can still be remembered in db, to work until it's replaced, and to be deleted.