Remove empty CC tracks Safari adds for HLS manifests without closed caption metadata
This is very odd and annoying -- we probably should be creating our HLS manifests with this metadata, but oddly AWS MediaConvert charges extra to do so! We could try to post-process the file to add it... or we can just remove teh weird tracks with no label specified, which seems to be what everyone else does.

  * https://github.com/videojs/video.js/issues/2808
  * https://developer.apple.com/library/archive/qa/qa1801/_index.html
  * https://github.com/samvera-labs/ramp/blob/23aae5c5aa9e7c95d1c94cba5cf870daf76df1aa/src/components/MediaPlayer/VideoJS/VideoJSPlayer.js#L571-L597