Reindex a parent work if an asset has it's WebVTT content or audio_asr_enabled flag changed
May have new or removed or changed VTT transcript content to be indexed.

We had an existing method of using ActiveRecord hooks to automatically trigger this parent work reindexes on asset changes. So I just extended it using the existing pattern. I don't necessarily love this pattern as it has been working out, but not the time to rethink it now.

After merging and deploying, a bulk reindex to ensure we're starting from consistency is a good idea.