Look at thumb not asset for height/width for aspect ratio
Normally the same, but when the asset is a PDF, the asset itself doesn't have a height/width, but it's thumb derivative (that we are about to display) does!  One part of the logic already was using `thumb`, it was just an error that the other part wasn't.

Makes it able to put aspect ratio in CSS for PDF thumbs when it couldn't before, avoiding (correct) error message, closing #3032.

I don't think it's worth adding a PDF test case, but maybe I'm wrong?