reload routes twice to reliably setup for tests
Fixes #2858.  I don't get it, and don't like that I don't understand it, and that it will make tests slighlty slower (maybe 10ms per test in this file).

But seems to work, good enough for now I think. Could be a bug in Rails or devise, not sure. Devise is known not to work quite right with Rails 8 (https://github.com/heartcombo/devise/pull/5728) so my guess is a devise bug