Pin sass to 1.77.6, to avoid so many deprecations for bootstrap
We upgraded sass to work with vite 7, https://github.com/sciencehistory/scihist_digicoll/pull/3157

But upgrading to latest sass gives us tons of deprecation warnings from bootstrap. Still works, just so many deprecation warnings (that we miss about our own code in the noise!).  Bootstrap does not seem to be in a hurry to fix, and fix won't be until a major version. https://github.com/twbs/bootstrap/issues/40849

sass 1.77.6 seems to be new enough to work with vite 7, but the last one that won't produce all those bootstrap deprecations. So we will lock to that. Wish we could put comments in package.json to explain why we're doing this!