Facets indexed to solr should be unicode normalized NFC
NFC normalization should not effect how text looks on screen -- but when we start using facet search/suggest feature, it can matter, as solr does not do any unicode normalization or analysis of 'string' fields we use for facetting, but needs to be a byte-by-byte match when using the facet filtering features the new UX feature has. Ref #3062.

So we will normalize NFC in our indexer before we send to solr.  Doesn't hurt to do immediately. Then in another PR, we can also ensure query is NFC normalized, so it can match.

Most of the stuff in our DB is probably NFC normalized already? But it can depend on exactly what a staff user entered/pasted into text entry for metadata entry!