Fix retrival of HLS playlist from ActiveEncode gem api, from MediaConvert
We use the ActiveEncode gem api to get back the status of the AWS MediaConvert async job.  When it's complete, we need to retrieve the URL to the created HLS playlist.

Either MediaConvert or the ActiveEncode API makes it hard to determine which of the outputs it lists is the main HLS playlist, vs subsidiary child playlists. We had been recognizing it based on it not having a `height` filled out -- but for whatever reason, it started having a height filled out recently.

Switch to recogizing it by it's name, just `hls.m3u8`, not `hls_low.m3u8`, or `hls_high.m3u8`, etc.