remove gem activerecord-postgres_enum, not needed since Rails 7
Gem for recording postgres enum types in schema.rb.  not needed since Rails 7.  That it was still there was resulting in double inclusion of the enum type in a dumped schema.rb, which we were assiduously avoiding committing, but it was annoying. Looks like this fixes it.

Slightly differnet form of statement in schema.rb, fixed now. Slightly different best practice in migration that uses it, fixed there too.